examples/Python_FrameBuffer_HelloWorld/requirements.txt
Pillow>=10.0
numpy>=1.24
evdev>=1.6
